# .well-known/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "User Management Plugin",
    "description": "A plugin that allows for basic user operations like adding, fetching, and updating users.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://scary-fashion-eel.cyclic.app"
    }
  ],
  "paths": {
    "/addUser": {
      "post": {
        "summary": "Add a new user",
        "description": "Create a new user with the supplied user details.",
        "operationId": "addUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User successfully added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/getUser/{userid}": {
      "get": {
        "summary": "Get details of a user",
        "description": "Fetches details of a user based on their userid.",
        "operationId": "getUser",
        "parameters": [
          {
            "name": "userid",
            "in": "path",
            "required": true,
            "description": "Userid of the user to be fetched",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/updateUser/{userid}": {
      "put": {
        "summary": "Update a user's details",
        "description": "Updates a user's details based on their userid.",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "userid",
            "in": "path",
            "required": true,
            "description": "Userid of the user to be updated",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/deleteUser/{userid}": {
      "delete": {
        "summary": "Delete a user",
        "description": "Deletes a user based on their userid.",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "userid",
            "in": "path",
            "required": true,
            "description": "Userid of the user to be deleted",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid ID supplied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "userid": {
            "type": "string",
            "description": "Unique identifier for the user"
          },
          "name": {
            "type": "string",
            "description": "Name of the user"
          },
          "email": {
            "type": "string",
            "description": "Email address of the user"
          },
          "password": {
            "type": "string",
            "description": "Password for the user account"
          }
        },
        "required": ["userid", "name", "email", "password"]
      }
    }
  }
}